wallet: type an amount at chain precision
send_amount_milli_eth was a u32 of thousandths of an ether with no
decimal point key, so the smallest amount expressible was 0.001. An exact
0.0005 or a single wei could not be entered at all, while staking and
swapping already took u128 with a point. The send screen had a cut down
third copy of an arithmetic that existed twice in better form.

One keypad now, shared. The figure is kept as typed, digits with a count
of places, and scaled only when the chain asks, so nothing rounds on the
way in and the amount signed is the amount shown.

sign_eth_transfer took value_wei as u64, capping a send at 18.4 ETH. The
wire already carries a 256 bit word built from a u128, so the narrowing
bought nothing and is gone.

24 host proofs: 0.0005 and one wei expressible, digits past the token's
precision refused rather than absorbed, backspace retracing the keys
pressed, and a max shortcut figure replaced by the next keystroke rather
than appended to.

+++ userland/capsule_wallet_nonos/src/wallet/ipc/sign_eth.rs
@@ -27,7 +27,7 @@ pub fn sign_eth_transfer(
     wallet_id: u32,
     to: [u8; 20],
     nonce: u64,
-    value_wei: u64,
+    value_wei: u128,
     gas_price_wei: u64,
 ) -> Result<Vec<u8>, i32> {
     let (max_priority, max_fee) = eip1559_fees(gas_price_wei);
@@ -40,7 +40,7 @@ pub fn sign_eth_transfer(
     push_word(&mut payload, max_fee);
     // A plain value transfer to an account always costs exactly 21000 gas.
     push_word(&mut payload, 21_000);
-    push_word(&mut payload, value_wei as u128);
+    push_word(&mut payload, value_wei);
     let rx = keyring_call(port, OP_SIGN_ETH_TRANSFER, &payload, 256)?;
     if rx.len() <= HDR_LEN {
         return Err(-11);
